requirements.txt
awesome-slugify==1.6.5
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-admin-jqueryui112==1.12.1b5
django-appconf==1.0.3
django-imagekit==4.0.2
django-phone-field==1.7.1
django-starcross-gallery==1.0.6
django-zip-code-api==1.4.2
idna==2.8
libmagic==1.0
pilkit==2.0
Pillow==5.0.0
psycopg2-binary==2.8.3
pytz==2019.1
regex==2019.6.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Unidecode==0.4.21
urllib3==1.25.3
